refactor: split db stores by persistence concern
