+++ test/db.test.js
@@ -6,7 +6,23 @@ import path from "node:path";
 
 import { runScan } from "../src/core/commands.js";
 import { loadConfig } from "../src/core/config.js";
-import { closeIndexDatabase, openIndexDatabase } from "../src/core/db.js";
+import { closeIndexDatabase, openIndexDatabase } from "../src/core/db/connection.js";
+import { getArtifact, listArtifacts, removeArtifact, upsertArtifact } from "../src/core/db/artifact-store.js";
+import {
+  loadFiles,
+  loadModules,
+  loadSymbols,
+  searchIndex,
+  writeRepositoryIndex,
+} from "../src/core/db/structural-store.js";
+import {
+  listSemanticProjects,
+  loadSemanticFileContext,
+  loadSemanticProjectFactsByFile,
+  loadSemanticRouteSurfaces,
+  replaceSemanticProjectSnapshot,
+  searchSemanticIndex,
+} from "../src/core/db/semantic-store.js";
 
 function octalMode(stats) {
   return (stats.mode & 0o777).toString(8);
@@ -45,4 +61,193 @@ test("openIndexDatabase read-only snapshots use user-only permissions", async ()
   }
 
   await assert.rejects(() => fs.access(tempDir));
+});
+
+test("structural store writes and searches repository index data", async () => {
+  const root = await fs.mkdtemp(path.join(os.tmpdir(), "agentify-db-structural-"));
+  const db = openIndexDatabase(root);
+  try {
+    writeRepositoryIndex(db, {
+      repo: {
+        name: "fixture",
+        root,
+        detected_stacks: ["ts"],
+        default_stack: "ts",
+        package_manager: "pnpm",
+      },
+      generated_at: "2026-04-27T00:00:00.000Z",
+      modules: [{
+        id: "app",
+        name: "App",
+        root_path: "src",
+        stack: "ts",
+        package_name: "fixture",
+        slug: "app",
+        doc_path: "docs/modules/app.md",
+        fingerprint: "module-fingerprint",
+        entry_files: ["src/index.ts"],
+        key_files: ["src/index.ts"],
+      }],
+      files: [
+        {
+          path: "src/index.ts",
+          module_id: "app",
+          language: "typescript",
+          size_bytes: 42,
+          fingerprint: "file-fingerprint",
+          is_test: 0,
+          is_config: 0,
+          is_entrypoint: 1,
+          is_key_file: 1,
+        },
+        {
+          path: "src/index.test.ts",
+          module_id: "app",
+          language: "typescript",
+          size_bytes: 21,
+          fingerprint: "test-fingerprint",
+          is_test: 1,
+          is_config: 0,
+          is_entrypoint: 0,
+          is_key_file: 0,
+        },
+      ],
+      symbols: [{
+        module_id: "app",
+        file_path: "src/index.ts",
+        name: "startApp",
+        kind: "function",
+        exported: 1,
+        start_line: 1,
+        end_line: 3,
+      }],
+      imports: [],
+      tests: [{
+        file_path: "src/index.test.ts",
+        module_id: "app",
+        framework: "node:test",
+        related_path: "src/index.ts",
+      }],
+      commands: [{
+        module_id: "app",
+        command_type: "test",
+        command: "pnpm",
+        args: ["test"],
+      }],
+    }, { headCommit: "abc123", provider: "local" });
+
+    assert.equal(loadModules(db)[0].id, "app");
+    assert.equal(loadFiles(db, "app").length, 2);
+    assert.equal(loadSymbols(db, "app")[0].name, "startApp");
+    assert.equal(searchIndex(db, "start").symbols[0].name, "startApp");
+  } finally {
+    closeIndexDatabase(db);
+  }
+});
+
+test("artifact store round-trips typed cached payloads", async () => {
+  const root = await fs.mkdtemp(path.join(os.tmpdir(), "agentify-db-artifacts-"));
+  const db = openIndexDatabase(root);
+  try {
+    upsertArtifact(db, {
+      key: "module-doc:app",
+      type: "module-doc",
+      scope: "app",
+      fingerprint: "fingerprint-1",
+      payload: { title: "App", sections: ["Overview"] },
+      updatedAt: "2026-04-27T00:00:00.000Z",
+    });
+
+    assert.deepEqual(getArtifact(db, "module-doc:app")?.payload, {
+      title: "App",
+      sections: ["Overview"],
+    });
+    assert.equal(listArtifacts(db, "module-doc").length, 1);
+
+    removeArtifact(db, "module-doc:app");
+    assert.equal(getArtifact(db, "module-doc:app"), null);
+  } finally {
+    closeIndexDatabase(db);
+  }
+});
+
+test("semantic store replaces snapshots and loads focused semantic facts", async () => {
+  const root = await fs.mkdtemp(path.join(os.tmpdir(), "agentify-db-semantic-"));
+  const db = openIndexDatabase(root);
+  try {
+    replaceSemanticProjectSnapshot(db, {
+      project: {
+        project_id: "tsconfig-json",
+        config_path: "tsconfig.json",
+        project_root: ".",
+        inferred: 0,
+        analyzer_version: "test",
+        schema_version: "semantic-v1",
+        status: "ready",
+        coverage_ratio: 1,
+        file_count: 1,
+        symbol_count: 1,
+        surface_count: 1,
+        edge_count: 1,
+        content_fingerprint: "content",
+        public_fingerprint: "public",
+        refreshed_at: "2026-04-27T00:00:00.000Z",
+      },
+      files: [{
+        project_id: "tsconfig-json",
+        file_path: "src/app/dashboard/page.tsx",
+        domain: "runtime",
+        is_header_target: 1,
+      }],
+      externalPackages: [{
+        project_id: "tsconfig-json",
+        package_name: "react",
+        usage_count: 1,
+      }],
+      symbols: [{
+        symbol_id: "symbol-dashboard",
+        project_id: "tsconfig-json",
+        file_path: "src/app/dashboard/page.tsx",
+        name: "DashboardPage",
+        display_name: "DashboardPage",
+        kind: "function",
+        export_name: "default",
+        start_line: 1,
+        end_line: 5,
+        is_exported: 1,
+        is_default: 1,
+        domain: "runtime",
+      }],
+      surfaces: [{
+        surface_id: "surface-dashboard",
+        project_id: "tsconfig-json",
+        file_path: "src/app/dashboard/page.tsx",
+        symbol_id: "symbol-dashboard",
+        kind: "route",
+        role: "page",
+        surface_key: "/dashboard",
+        display_name: "Dashboard",
+        domain: "runtime",
+        is_header_target: 1,
+      }],
+      symbolEdges: [{
+        project_id: "tsconfig-json",
+        from_symbol_id: "symbol-dashboard",
+        from_file_path: "src/app/dashboard/page.tsx",
+        to_external_package: "react",
+        edge_kind: "import",
+        edge_domain: "runtime",
+        confidence: 1,
+        source: "test",
+      }],
+    });
+
+    assert.equal(listSemanticProjects(db)[0].status, "ready");
+    assert.equal(loadSemanticRouteSurfaces(db)[0].surface_key, "/dashboard");
+    assert.equal(loadSemanticProjectFactsByFile(db)[0].surface.displayName, "Dashboard");
+    assert.equal(loadSemanticFileContext(db, "src/app/dashboard/page.tsx").exports[0].export_name, "default");
+    assert.equal(searchSemanticIndex(db, "dashboard").semantic_surfaces[0].surface_key, "/dashboard");
+  } finally {
+    closeIndexDatabase(db);
+  }
 });
